move scm endpoints into separate controller, change partner endpoints cli back to normal
