+++ python/sdk/prelude_sdk/controllers/partner_controller.py
@@ -65,24 +65,7 @@ class PartnerController(HttpController):
         if res.status_code == 200:
             return res.json()
         raise Exception(res.text)
-    
-    @verify_credentials
-    def endpoints_via_scm(self, partner: Control, filter: str, orderby: str, top: int):
-        """ Get a list of endpoints from a partner with SCM analysis """
-        params = {
-            '$filter': filter,
-            '$orderby': orderby,
-            '$top': top
-        }
-        res = self._session.get(
-            f'{self.account.hq}/partner/endpoints/{partner.name}',
-            headers=self.account.headers,
-            params=params,
-            timeout=30
-        )
-        if res.status_code == 200:
-            return res.json()
-        raise Exception(res.text)
+
 
     @verify_credentials
     def generate_webhook(self, partner: Control):
@@ -172,42 +155,6 @@ class PartnerController(HttpController):
             headers=self.account.headers,
             params=params,
             timeout=30
-        )
-        if res.status_code == 200:
-            return res.json()
-        raise Exception(res.text)
-
-    @verify_credentials
-    def get_policy_evaluation_summary(self):
-        """ Get policy evaluation summary for all partners """
-        res = self._session.get(
-            f'{self.account.hq}/partner/policy_evaluation_summary',
-            headers=self.account.headers,
-            timeout=30
-        )
-        if res.status_code == 200:
-            return res.json()
-        raise Exception(res.text)
-
-    @verify_credentials
-    def get_policy_evaluation(self, partner: Control):
-        """ Get policy evaluations for given partner """
-        res = self._session.get(
-            f'{self.account.hq}/partner/policy_evaluation/{partner.name}',
-            headers=self.account.headers,
-            timeout=30
-        )
-        if res.status_code == 200:
-            return res.json()
-        raise Exception(res.text)
-
-    @verify_credentials
-    def update_policy_evaluation(self, partner: Control):
-        """ Update policy evaluations for given partner """
-        res = self._session.post(
-            f'{self.account.hq}/partner/policy_evaluation/{partner.name}',
-            headers=self.account.headers,
-            timeout=60
         )
         if res.status_code == 200:
             return res.json()
